+++ mipengine/common/resources/pathologies_metadata/dementia.json
