+++ mipengine/common/resources/pathologies_metadata/mentalhealth.json
