+++ mipengine/node/config/config.ini
@@ -12,5 +12,5 @@ username = monetdb
 password = monetdb
 
 [node]
-identifier = global_node
+identifier = local_node_1
