+++ mipengine/node/monetdb_interface/remote_tables.py
@@ -5,7 +5,7 @@ from mipengine.node.monetdb_interface.common import connection
 from mipengine.node.monetdb_interface.common import convert_schema_to_sql_query_format
 from mipengine.node.monetdb_interface.common import cursor
 from mipengine.node.tasks.data_classes import TableInfo
-from mipengine.utils.validate_identifier_names import validate_identifier_names
+from mipengine.common.validate_identifier_names import validate_identifier_names
 
 
 def get_remote_tables_names(context_id: str) -> List[str]:
