+++ requirements/tests.txt
@@ -1,3 +1,4 @@
 pytest~=6.2.2
 pytest-xdist
 requests~=2.22.0
+pandas
